WFP-2684 logging, changed exception and move delete method
