+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsallocations/listener/OffenderEventListener.kt
@@ -11,12 +11,14 @@ import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 import uk.gov.justice.digital.hmpps.hmppsallocations.client.EventsNotFoundError
 import uk.gov.justice.digital.hmpps.hmppsallocations.client.ForbiddenOffenderError
+import uk.gov.justice.digital.hmpps.hmppsallocations.service.UnallocatedDataBaseOperationService
 import uk.gov.justice.digital.hmpps.hmppsallocations.service.UpsertUnallocatedCaseService
 
 @Component
 class OffenderEventListener(
   private val objectMapper: ObjectMapper,
   private val upsertUnallocatedCaseService: UpsertUnallocatedCaseService,
+  private val unallocatedDataBaseOperationService: UnallocatedDataBaseOperationService,
 ) {
 
   @SqsListener("hmppsoffenderqueue", factory = "hmppsQueueContainerFactoryProxy")
@@ -30,7 +32,7 @@ class OffenderEventListener(
         log.warn("Unable to access offender with CRN $crn with error: ${e.message}")
       } catch (e: EventsNotFoundError) {
         log.warn("Unable to find events for CRN $crn with error: ${e.message}")
-        upsertUnallocatedCaseService.deleteEventsForNoActiveEvents(crn)
+        unallocatedDataBaseOperationService.deleteEventsForNoActiveEvents(crn)
       }
     }.get()
   }
